Enhance the dynamic type support for BasicInfoCell
Before the app was waiting for the user to change the larger accessibility fonts
and then do the necessary UI updates.

This behavior was and still is right, but what if the user already had the larger accessibility fonts applied
before opening the app! The current behavior is to have 2 UI states, and apply it accordingly either at initialization
or if the user change the accessibility fonts while using the app